Added mechanism to provide a version to shlink-web-client
